requirements.txt
djangorestframework-simplejwt
djangorestframework
psycopg2-binary
django-redis
Django>=5.0
gunicorn
requests
redis